backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn>=0.41.0",
    "passlib[bcrypt]>=1.7.4",
]
